LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/canvas/offscreen/manual/the-offscreen-canvas/offscreencanvas.getcontext-expected.txt


PASS Test that getContext with un-supported string throws a TypeError.
FAIL Test that getContext with supported string returns correct results assert_true: expected true got false
FAIL Test that getContext twice with different context type returns null the second time assert_equals: expected null but got object "[object OffscreenCanvasRenderingContext2D]"
PASS Test that 2dcontext.canvas should return the original OffscreenCanvas
FAIL Test that webglcontext.canvas should return the original OffscreenCanvas null is not an object (evaluating 'ctx.canvas')
FAIL Test that OffscreenCanvasRenderingContext2D with alpha disabled makes the OffscreenCanvas opaque assert_approx_equals: Green channel of the pixel at (5, 5) expected 127 +/- 2 but got 255
PASS Test that OffscreenCanvasRenderingContext2D with alpha enabled preserves the alpha
PASS Test that 'alpha' context creation attribute is true by default
